refactor(#2453 R2 R3): re-home commit-pending witnesses to a sibling *tests* file
CI `src_file_size_invariant` (a FULL-nextest gate, not caught by clippy or the
"LOC overrun" check) flagged app/mod.rs at 3511 LOC — 54 over its grandfathered
anti-monolith ceiling (3457) — because R3 added the poll_commit_pending witnesses
inline. Move them to a sibling src/app/commit_pending_tests.rs (exempt by filename);
`use super::*` reaches app's private CommitPending / CommitPoll / poll_commit_pending.
app/mod.rs → 3443 (< 3457). No production change; the 5 witnesses pass unchanged from
the new home.


Agend-Agent: archfix-claude-opus
Agend-Task: t-20260712034222302055-19288-5
Agend-Branch: feat/app-staged-restart-r2
Agend-Issued-At: 2026-07-12T03:43:18.424715+00:00
